handler: no longer an object but a union of action table id and object id - handler execute is a little faster

+++ include/ui/handler.h
@@ -3,6 +3,6 @@
 #ifndef _HANDLER_H_
   #define _HANDLER_H_
 
-  short Handler_create(const short destination_id, const short table_id);
-  void Handler_execute(const short handler_id, const short slot_id, const short source_id);
+  int Handler_create(const short destination_id, const short table_id);
+  void Handler_execute(const int handler_id, const short slot_id, const short source_id);
 #endif
