+++ include/ui/ui_text.h
@@ -5,6 +5,6 @@
 
   short UI_text_create(void);
   short UI_text_text_id(const short id);
-  short UI_text_area_handler(const short ui_text_id);
+  int UI_text_area_handler(const short ui_text_id);
   void UI_text_add(const short id, const char *str);
 #endif
